+++ src/containerapp/HISTORY.rst
@@ -8,6 +8,8 @@ upcoming
 * 'az containerapp function keys show': Show specific function key in a container app
 * 'az containerapp function keys list': List function keys in a container app
 * 'az containerapp function keys set': Set/Update specific function key in a container app
+* 'az containerapp function invocations summary': Get function invocation summary from Application Insights
+* 'az containerapp function invocations traces': Get function invocation traces from Application Insights
 * 'az containerapp update/up': Disallow changing `--revisions-mode` to Labels.
 * 'az containerapp session code-interpreter': Fix `--path` in examples
 * 'az containerapp sessionpool create/update': Support `--lifecycle-type` and `--max-alive-period`